move fetching to separate api.ts file
