+++ assets/src/components/app.tsx
@@ -1,10 +1,7 @@
 import React, { useEffect, useReducer } from "react"
+import * as Api from "../api"
 import { Route } from "../skate.d"
 import RoutePicker from "./route_picker"
-
-interface RoutesRequestResponse {
-  data: Route[]
-}
 
 interface State {
   routes: null | Route[]
@@ -44,20 +41,10 @@ const App = (): JSX.Element => {
   const [state, dispatch] = useReducer(reducer, initialState)
 
   useEffect(() => {
-    fetch("/api/routes")
-      .then((response: Response) => {
-        if (response.status !== 200) {
-          throw new Error(`Response error: ${response.status}`)
-        }
-        return response
-      })
-      .then((response: Response) => response.json())
-      .then(({ data: routes }: RoutesRequestResponse) =>
-        dispatch(setRoutes(routes)),
+    Api.fetchRoutes()
+      .then((routes: Route[]) =>
+        dispatch(setRoutes(routes))
       )
-      .catch(error => {
-        console.log(`Error loading routes`, error)
-      })
   }, [])
 
   return (
